+++ src/main/java/Albaid/backend/domain/card/application/AlbaCardService.java
@@ -3,8 +3,8 @@ package Albaid.backend.domain.card.application;
 import Albaid.backend.domain.card.application.dto.AlbaCardDTO;
 
 public interface AlbaCardService {
-    AlbaCardDTO getAlbaCardById(Long id);
+    AlbaCardDTO getAlbaCardById(Integer id);
     AlbaCardDTO createAlbaCard(AlbaCardDTO albaCardDto);
-    AlbaCardDTO updateAlbaCard(Long id, AlbaCardDTO albaCardDto);
-    void deleteAlbaCard(Long id);
+    AlbaCardDTO updateAlbaCard(Integer id, AlbaCardDTO albaCardDto);
+    void deleteAlbaCard(Integer id);
 }
