+++ src/main/java/Albaid/backend/domain/card/repository/AlbaCardRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AlbaCardRepository extends JpaRepository<AlbaCard, Long> {
+public interface AlbaCardRepository extends JpaRepository<AlbaCard, Integer> {
 }
